chore: Move NavigationListener component into its own file
This removes the warning when running the linter
